rss2email: add ability to load feeds from the config file only.
If a feed exists in the config file, but not in the data file, the
previous implementation would not load it.  This patch initializes
such feeds using only the information from the config file
(i.e. without dynamic data from the data file).  The patch also
creates missing data files on demand.  As an example use case, if you
keep a backup of your config file, but lose the data file, you can
restore the config file and an `r2e run` will create and repopulate
your data file.