test:gmane: add Gmane feed for RSS testing
The weird indentation is because Gmane wraps the descriptions in <pre>
tags (to preserve formatting in the initial email).  html2text is
converting the description to Markdown, so it inserts a leading 4
spaces for preformatted blocks.  The unindented initial line and
following blank are due to a bug in html2text, for which I've
submitted

  https://github.com/aaronsw/html2text/pull/63