CONTRIBUTING: point to external docs (for non GPLv2= projects)
If a project wants to use the DCO/s-o-b workflow, but can't because of
license incompatibility, they can include this blurb pointing towards
the external documentation.  To avoid licensing issues with this
CONTRIBUTING file itself, I'm releasing it under the Creative Commons
CC0 1.0 Universal license (see the `license` branch for full text).

If you use this in your project, you'll probably want to adjust the
URL to point to somewhere more dependable.  It's annoyingly long, but
you should include the blob hash (or a higher level hash like the
commit hash) in your URL to make it absolutely clear which version of
the documentation you were using in your project at a particular time.

Signed-off-by: W. Trevor King <wking@tremily.us>