Drop smtp-ssl-protocol config option (part 1)
Selecting the highest available ssl/tls version is always reasonable (I can't think of a single situation where it shouldn't be used) and this simplifies email.py, as well as the effort users need to put into configuring smtp in rss2email.cfg, as they will have to research what SSLv23 actually means. (It's not straight forward that this also enabled TLS).