User agent needs to be substituted at use time, not at load time
If we substitute `__VERSION__` and `__URL__` in the user agent string
when it is loaded from the configuration file, then whenever a new
feed is added it will end up with the current version number and URL
hard-coded into its `user-agent` setting, which means that even when
either of those changes in a new version of rss2email, that feed will
keep using the old ones.

This fix changes that by doing the substitution at the time the user
agent string is actually used, i.e., included in an HTTP request or
email header, rather than when the config file is loaded.