Fix per-feed user-agent settings broken by user agent change in 3.11
Because we were substituting `__VERSION__` and `__URL__` into the user
agent string when the config was read in 3.11, any feed added with
3.11 would end up with a per-feed user-agent setting with substituted
current values, so that feed would keep using those values even when
the version or URL of rss2email changed. That bug is fixed in another
commit, but now we want to put in some migration code to repair feeds
that were broken in the interim. We can't get this 100% right since we
can only safely do the repair when the user is using the default
user-agent setting, but that will cover most users so it is worth
doing.