Use `parameterized` for tests, fix nondeterministic required `chdir`
Use the `parameterized` module to treat each `.config` file as a
separate test, so that it's clear which of them is failing, and so
that it is possible to use the functionality built into `unittest` for
running subsets of the tests or even individual tests.

While I was implementing this I discovered that all of the test cases
were depending on a call to `_os.chdir` being done in just one of
them, which is problematic if you're only trying to execute a subset
of the tests and also because it's a bad idea to rely on the order in
which tests execute, so I moved the `_os.chdir` into `ExecContext` and
made sure all the code which depends on the current directory being
changed is wrapped within an `ExecContext`.

This commit also puts code into `test.py` for adjusting the python
search path as necessary so that the user calling the tests no longer
needs to do that, and updates `tests/README` to reflect that change
and update the instructions for running a subset of the tests to
reflect that they are now being executed by the `unittest` framework.