Fix html2text config options
The three html2text options have never worked due to the way they were
set. In addition, the body width in html2text defaults to 78, and the
default of zero in r2e should have disabled wrapping, but was instead
wrapping at 78.

This change uses negative values to disable wrapping, and for
compatibility reasons, the default of zero will continue wrapping at 78.