Log sendmail output
There's really no way to debug a broken sendmail setup when the output
is thrown away. Stdout and stderr were passed to the exception as
attributes but do not appear to be used there.

I don't believe there's any other good way to merge stdout/stderr and
preserve the sequential order than to pass one to the other and the
other to PIPE.