Fix double escaping of sender names on certain configurations (#166)
Before this patch, the sender name and address went through a
formataddr→parseaddr→formataddr series of transformations, with a
Header(…, guess_encoding(…)).encode() of the name part before the final
formataddr. This produces a correct message, so long as 'US-ASCII' appears first
in the user-configurable list of encodings for guess_encoding.

It is important to note that the first formataddr, in the series of
transformations above, already encodes the name part as necessary to produce a
US-ASCII string. Therefore, if the sender name includes non-ASCII characters and
guess_encoding does not return 'US-ASCII', then the Header().encode() step would
encapsulate the encoded string that was already encapsulated by the first
formataddr, thereby producing an erroneous double escaping. The MUA, in this
case, will dutifully display an unfriendly From: =?utf-8?b?…?= <…@…>.

The fix is to let formataddr do its job and not to re-process its result. Also
note that formataddr, in addition to optionally encoding the name part, makes
sure that the address part is pure ASCII.

The patch also prepends 'US-ASCII' to the list of user-provided encodings, for the
encoding of all other headers. While not strictly necessary in order to produce
a correctly encoded message, it does make the raw message easier to work with,
and helps to avoid modifying some parts of the TestEmails class in non-trivial
ways.

The included new test case demonstrates nicely how the badly-encoded message
looks like, when run with the old code.