Fix status parsing of file:// feeds on Python>=3.9
Starting with Python 3.9, the return value from `urllib.request.urlopen` changed
to always have the `.status` attribute, becoming `None` in cases where it
previously was absent (notably on file:// URIs).

Starting with feedparser version 6.0.0b1, it attempts to hide this detail on the
return value from `feedparser.parse` by setting `.status` to `200` in such
cases, but on Python 3.9 the `None` value slips through nonetheless.

A later release of feedparser may fix the issue and successfully hide this
detail, but in the meantime we can support the already-released 6.0.x versions
on all current Python versions by explicitly ignoring the `None` value.