Change "entry" and "body" from html id to class.
Unlike header and footer elements, entry and body were used as ids,
not class.  Not only was this inconsistent, it also didn't actually
work, because the "#entry" and "#body" lines in the default config
file were parsed as comments.

For backward compatibility with users who have defined their own
style sheets, we are keeping them as ids as well
(e.g., <div class="body" id="body">).