Fix a couple of HTTP user agent regressions
At the same time:
* Go back to using `rss2email` in the UA, after erroneously using `feedparser`.
* Regain the ability to override the HTTP user agent.

These represent a breaking change for users who might have come to rely on the
incorrect behaviour since version 3.11, so the release notes have been updated
with guidance for those affected.

The entry in the manual for `user-agent` configuration was moved to the
miscellaneous section because it affects both fetching feeds and sending mails.

It is worth noting that f57513a (Use configured user-agent string and replace
_VERSION_ with __version__ value, 2019-09-13) changed the semantics of the
`user-agent` configuration unintentionally. This error was unfortunately added
to the documentation in 574ba59 (Add user-agent documentation to man page,
2019-09-13), causing further confusion.

Co-authored-by: Eric Durand <rand0mdud31@gmail.com>