support oauth2 login
Initial support for oauth2 login.
The user needs to set 4 variables in the config file:

imap-oauth2 is a boolean that needs to be set to True.
imap-clientid, imap-clientsecret, imap-refreshtoken are needed to log in.

The new file oauth2.py also includes code to generate a refresh
token. Currently the function needs to be called manually.